Fix Long/Int comparison in VaultCtrDataSource
dataSpec.length is Long; C.LENGTH_UNSET is Int. Coerce with .toLong().

+++ app/src/main/java/com/atelierapps/vault/crypto/VaultCtrDataSource.kt
@@ -64,7 +64,7 @@ class VaultCtrDataSource(private val context: Context) : BaseDataSource(/* isNet
         raf = RandomAccessFile(file, "r").apply { seek(headerLen + position) }
         val dataLength = file.length() - headerLen
         bytesRemaining =
-            if (dataSpec.length != C.LENGTH_UNSET) dataSpec.length else dataLength - position
+            if (dataSpec.length != C.LENGTH_UNSET.toLong()) dataSpec.length else dataLength - position
 
         opened = true
         transferStarted(dataSpec)
